Hide buttons that the user can't click issue #75
Resolves issue #75
Instead of displaying an error message, I just hide the buttons when the default tweet is shown. Then, when a user clicks a tweet, the buttons to like, follow, etc are unhidden.
This is probably a better solution, since clicking those buttons will never do anything. They might as well not be there